# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sheets/snapshot:
    get:
      operationId: getSheetsSnapshot
      tags: [sheets]
      summary: Retrieve the application snapshot from Google Sheets
      responses:
        "200":
          description: Snapshot data organized by worksheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetsSnapshot"
        "409":
          description: Google Sheets is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sheets/configure:
    post:
      operationId: configureSheets
      tags: [sheets]
      summary: Configure the target Google Sheets workbook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SheetsConfigInput"
      responses:
        "200":
          description: Saved Google Sheets configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetsStatus"
  /sheets/initialize:
    post:
      operationId: initializeSheets
      tags: [sheets]
      summary: Create or initialize the application workbook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SheetsInitializeInput"
      responses:
        "200":
          description: Initialized Google Sheets workbook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetsStatus"
        "409":
          description: Google Sheets write access is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sheets/sync:
    post:
      operationId: syncSheets
      tags: [sheets]
      summary: Push application data to Google Sheets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SheetsSyncInput"
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetsSyncResult"
        "409":
          description: Google Sheets write access is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    JsonValue:
      nullable: true
      oneOf:
        - type: string
        - type: number
        - type: boolean
        - type: object
          additionalProperties: true
        - type: array
          items: {}
    SheetRow:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/JsonValue"
    SheetTab:
      type: object
      required: [name, rows]
      properties:
        name:
          type: string
          minLength: 1
        rows:
          type: array
          items:
            $ref: "#/components/schemas/SheetRow"
    SheetsStatus:
      type: object
      required: [configured, writable, connection, accessState, lastSyncedAt, spreadsheetId, spreadsheetUrl]
      properties:
        configured:
          type: boolean
        writable:
          type: boolean
        connection:
          type: string
        accessState:
          type: string
          enum: [connected, not_configured, provider_error]
        errorCode:
          type: ["string", "null"]
        message:
          type: string
        lastSyncedAt:
          type: ["string", "null"]
        spreadsheetId:
          type: ["string", "null"]
        spreadsheetUrl:
          type: ["string", "null"]
    SheetsSnapshot:
      type: object
      required: [status, tabs, revision]
      properties:
        status:
          $ref: "#/components/schemas/SheetsStatus"
        tabs:
          type: array
          items:
            $ref: "#/components/schemas/SheetTab"
        revision:
          type: string
          minLength: 1
        canonicalIndex:
          type: object
          additionalProperties: true
    SheetsConfigInput:
      type: object
      properties:
        spreadsheetId:
          type: string
          minLength: 10
        spreadsheetUrl:
          type: string
        title:
          type: string
          minLength: 1
      anyOf:
        - required: [spreadsheetId]
        - required: [spreadsheetUrl]
    SheetsInitializeInput:
      type: object
      required: [title, tabs]
      properties:
        title:
          type: string
          minLength: 1
        spreadsheetId:
          type: string
          minLength: 10
        tabs:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SheetTab"
    SheetsSyncInput:
      type: object
      required: [tabs, expectedRevision]
      properties:
        tabs:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SheetTab"
        expectedRevision:
          type: string
          minLength: 1
    SheetsSyncResult:
      type: object
      required: [status, updatedTabs, updatedRows, revision]
      properties:
        status:
          $ref: "#/components/schemas/SheetsStatus"
        updatedTabs:
          type: number
        updatedRows:
          type: number
        revision:
          type: string
          minLength: 1
        canonicalIndex:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        category:
          type: string
        recoverable:
          type: boolean
...